requirements.txt
aiogram~=3.19.0
SQLAlchemy~=2.0.40
# dotenv~=0.9.9
python-dotenv~=1.1.0
